pyproject.toml
[tool.poetry]
name = "scrapersky"
version = "0.1.0"
description = ""
authors = ["wjbradshawTerp <157817565+wjbradshawTerp@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
xclienttransaction = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
